pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "things-mcp"
version = "0.1.0"
description = "Things app Model Context Protocol (MCP) server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.0",
    "things-py>=0.0.15",
]

[project.scripts]
things-mcp = "things_mcp.__main__:main"

[tool.setuptools]
packages = ["things_mcp"]
package-dir = {"" = "src"}
